remove unnecessary components
