+++ stockTrainerFrontend/src/components/billing/Billing.jsx
@@ -10,8 +10,6 @@ const UpgradeUser = props => {
       <div className="upgradeInfo">
         <h3>Upgrade to StockTrainer Premium</h3>
         <p>Keep track of all your favorite stocks</p>
-        <p>Look at all 5 of our fancy indicators</p>
-        <p>We help you find stocks</p>
       </div>
       {/* make sure to have your stripe api key set */}
       <StripeProvider
